Crate vertex

Modules

errors

Module containing custom error types for the project.

graph_utils

Module containing functions to manipulate graphs used in the project.

maxsat

Module containing the structures and algorithms for the satLB lower bound of the BnB algorithm. This module can only be used in this context. Using the methods outside this algorithm may lead to unexpected results.

result_utils

Module containing tools to handle the results of the algorithms. (Time taken, minimum vertex cover, etc.)

Functions

branch_and_bound

Branch and bound algorithm that searches for the minimum vertex cover of a given graph.

naive_search

Naïve algorithm that searches for the minimum vertex cover of a given graph.

numvc

Local search based algorithm that search for the minimum vertex cover of a given graph.

read_arguments

Read the command line arguments given as parameters and run the algorithm based on the given arguments.

run_algorithm

Run a given algorithm on a given graph and print the result.

samvc

Simulated Annealing based algorithm that search for the minimum vertex cover of a given graph.