Function vertex::branch_and_bound

pub fn branch_and_bound(
    graph: &UnGraphMap<u64, ()>,
    clock: &mut Clock,
    _params: Option<&[f64]>,
    optimal: Option<u64>
) -> (u64, Vec<u64>)

Branch and bound algorithm that searches for the minimum vertex cover of a given graph.

Branch : The algorithm branches on the vertex with max degree. It creates 2 branches : one where the vertex is in the vertex cover and one where its neighbors are in the vertex cover.

Bound : The algorithm has 2 lower bounds : clqLB and degLB. (see the paper linked in README for more details)

The clock is used to stop the algorithm if it reaches the time limit defined in the clock. It is also used to measure the time taken by the algorithm (and some of its subroutines).

Example

use petgraph::prelude::UnGraphMap;
use vertex::branch_and_bound;
use vertex::graph_utils::load_clq_file;
use vertex::result_utils::Clock;

let graph = load_clq_file("src/resources/graphs/test.clq")
            .expect("Error while loading graph");
let mut clock = Clock::new(3600); // 1 hour time limit

let res = branch_and_bound(&graph, &mut clock, None, None);

assert_eq!(res.0, 3);
assert_eq!(res.1, vec![0, 4, 2]);