Function vertex::run_algorithm

pub fn run_algorithm(
    graph_id: &str,
    graph: &UnGraphMap<u64, ()>,
    f: &dyn Fn(&UnGraphMap<u64, ()>, &mut Clock) -> (u64, Vec<u64>),
    cmpl: bool
) -> Result<MVCResult, YamlError>

Run a given algorithm on a given graph and print the result.

It is the default function when you want to test your algorithm on a certain graph. It prints the result and tell you if it is optimal or not based on the data in the yaml file. The algorithm must take an UnGraphMap as input and returns u64.

Example

use petgraph::prelude::UnGraphMap;use vertex::graph_utils::load_clq_file;
use vertex::{naive_search, run_algorithm};

let mut graph = load_clq_file("src/resources/graphs/test.clq").unwrap();
let res = run_algorithm("test.clq", &graph, &naive_search, false).unwrap_or_else(|e| {
   panic!("Error while running algorithm : {}", e);
});
println!("{}", res);