Module vertex::graph_utils

Module containing functions to manipulate graphs used in the project.

Structs

GraphInfo

Structure used to store the information of a graph such as its exact value of the MVC.

YamlTime

Structure used to store the information of a computation of the MVC for a given graph.

Functions

add_graph_to_yaml

Add the graph id with its format in the yaml file located at src/resources/graph_data.yml.

add_time_to_yaml

Adds a new time for the given graph to the yaml file located at src/resources/time_result.yml.

complement

Returns the complement of a given graph.

copy_graph

Since clone is not implemented for MatrixGraph, this function manually copies the graph. It iterates over the nodes and edges of the graph and adds them to the copy.

get_optimal_value

Get the optimal value for a given graph id. The optimal value is the value stored in the yaml file. So, if the value in the yaml file is wrong, this function will return the wrong result.

get_time_data

Get all the times for a given graph id.

get_vertex_with_max_degree

Returns the vertex with the maximum degree in the graph and its degree.

graph_to_string

Returns the string of a given file in the DIMACS .clq format.

is_clique

Check if a given array of vertices is a clique in the given graph.

is_independent_set

Check if a given array of vertices is an independent set in the given graph.

is_optimal_value

Check if a given value is the optimal value for a given graph id. If the graph id is not in the file, return None.

is_vertex_cover

Check if a given vertex cover is a vertex cover of a given graph.

load_clq_file

Load a graph from a DIMACS .col file.

update_mvc_value

Update the known value of the minimum vertex cover for a given graph id.