Function vertex::graph_utils::is_optimal_value

pub fn is_optimal_value(
    id: &str,
    val: u64,
    path: Option<&str>
) -> Result<Option<bool>, YamlError>

Check if a given value is the optimal value for a given graph id. If the graph id is not in the file, return None.

The optimal value is the value stored in the yaml file. So, if the value in the yaml file is wrong, this function will return the wrong result.

Parameters

id : the id of the graph (ex: test.clq)

val : the value to check

path : the path to the yaml file containing the graph info (optional-> None or Some(path))

Throws

YamlError::IoError if the file cannot be opened or is not found

YamlError::YAMLParsingError if there is an error while parsing the file

example

use vertex::graph_utils::is_optimal_value;

let true_res = is_optimal_value("test.clq", 3, Some("src/resources/graph_data.yml")).unwrap_or_else(|e| {
   panic!("Error while checking if value is optimal : {}", e);
});
assert!(true_res.unwrap());

let false_res = is_optimal_value("test.clq", 2, Some("src/resources/graph_data.yml")).unwrap_or_else(|e| {
panic!("Error while checking if value is optimal : {}", e);
});
assert!(!false_res.unwrap());