Struct vertex::Clock

pub struct Clock {
    pub start: Instant,
    /* private fields */
}

Struct representing a clock used to measure the time taken by an algorithm and stop it if it reaches the time limit.

This clock is based on the std::time::Instant struct. The clock can also be used to measure the time taken by some subroutines of the algorithm.

Example

use std::time::Duration;
use vertex::Clock;

let mut clock = Clock::new(3600); // 1 hour time limit

let elapsed = clock.get_time();

clock.enter_subroutine("subroutine1");
// Do something
clock.exit_subroutine("subroutine1").expect("The subroutine was not entered before");
clock.enter_subroutine("subroutine2");
// Do something
clock.exit_subroutine("subroutine2").expect("The subroutine was not entered before");

if clock.is_time_up() {
   println!("Time is up !");
}
println!("Time taken by the algorithm : {}", elapsed);
println!("Time taken by subroutine1 : {}", clock.get_subroutine_duration("subroutine1").as_millis());

Fields
start: Instant
Implementations

impl Clock

pub fn new(limit: u64) -> Clock

pub fn get_time(&self) -> ElapseTime

Returns the time off the clock since it was created.

pub fn stop_timer(&mut self)

Stops the clock.

pub fn is_time_up(&self) -> bool

Returns true if the time limit is reached.

pub fn enter_subroutine(&mut self, name: &str)

Enters a subroutine of the algorithm and start the timer for this subroutine. It creates a new start time for this subroutine but don’t reset the duration.

If the subroutine was already entered before, it will reset the start time and add the time taken since the last time it was entered.

Example

use std::time::Duration;
use vertex::Clock;

let mut clock = Clock::new(3600);

clock.enter_subroutine("subroutine1");
// Do something
clock.enter_subroutine("subroutine2");
// Do something
clock.exit_subroutine("subroutine2").expect("The subroutine was not entered before");

clock.enter_subroutine("subroutine1");
// Add the time taken since the last time we entered subroutine1
clock.exit_subroutine("subroutine1").expect("The subroutine was not entered before");

pub fn exit_subroutine(&mut self, name: &str) -> Result<(), ClockError>

Exits a subroutine of the algorithm and add the time taken since the last time it was entered. If the subroutine was already exit before, it does nothing.

Throws

ClockError if the subroutine was not entered before.

pub fn get_subroutine_duration(&self, name: &str) -> Duration

Returns the time taken by a subroutine of the algorithm.

The time taken is the sum of all the time taken by this subroutine since the first time it was entered. If the subroutine was not entered before, it a duration of 0.

Example

use std::time::Duration;
use vertex::{Clock, ElapseTime};

let mut clock = Clock::new(3600);

clock.enter_subroutine("subroutine1");
// Do something
clock.exit_subroutine("subroutine1").expect("The subroutine was not entered before");

let elapsed = clock.get_subroutine_duration("subroutine1");
println!("Time taken by subroutine1 : {}", ElapseTime::new(elapsed));
println!("Percentage of time taken by subroutine1 : {}%", elapsed.as_secs_f64() * 100.0 / clock.get_time().duration.as_secs_f64());